+++ tests/TestApp/app/Model/ApplicationModel.php
@@ -6,13 +6,18 @@ use Origin\Model\Model;
 
 class ApplicationModel extends Model
 {
+    protected function initialize(): void
+    {
+        $this->onError('errorHandler');
+    }
+    
     /**
      * This is callback is called when an exception is caught
      *
      * @param \Exception $exception
      * @return void
      */
-    public function onError(\Exception $exception): void
+    public function errorHandler(\Exception $exception): void
     {
     }
 }